uploader/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile uploader/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
cachetools==6.2.6
    # via pyiceberg
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via pyiceberg
colorama==0.4.6
    # via click
cryptography==46.0.5
    # via pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
fsspec==2026.2.0
    # via pyiceberg
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via pyiceberg
multidict==6.7.1
    # via yarl
packaging==26.0
    # via deprecation
pillow==12.1.1
    # via -r uploader/requirements.in
postgrest==2.28.0
    # via supabase
propcache==0.4.1
    # via yarl
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   postgrest
    #   pyiceberg
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyiceberg==0.11.0
    # via storage3
pyjwt[crypto]==2.11.0
    # via supabase-auth
pyparsing==3.3.2
    # via pyiceberg
pyroaring==1.0.3
    # via pyiceberg
python-dateutil==2.9.0.post0
    # via strictyaml
python-dotenv==1.2.2
    # via -r uploader/requirements.in
python-slugify==8.0.4
    # via -r uploader/requirements.in
realtime==2.28.0
    # via supabase
requests==2.32.5
    # via pyiceberg
rich==14.3.3
    # via pyiceberg
six==1.17.0
    # via python-dateutil
storage3==2.28.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
strictyaml==1.7.3
    # via pyiceberg
supabase==2.28.0
    # via -r uploader/requirements.in
supabase-auth==2.28.0
    # via supabase
supabase-functions==2.28.0
    # via supabase
tenacity==9.1.4
    # via pyiceberg
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
websockets==15.0.1
    # via realtime
yarl==1.23.0
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-functions
zstandard==0.25.0
    # via pyiceberg
